+++ app/Http/Controllers/FotoController.php
@@ -6,6 +6,8 @@ use App\Models\Foto;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
+use App\Models\Evento;
+use Illuminate\Support\Facades\Gate;
 class FotoController extends Controller
 {
     /**
@@ -27,7 +29,7 @@ class FotoController extends Controller
         //Fotografia_carrusel
         $foto=Foto::create($request->except('_token','_method','Fotografia_carrusel'));
         $evento=$foto->evento;
-        return redirect()->route('evento.edit',$evento);
+        return redirect()->route('evento.edit',$evento)->with('success','Se agrego una foto nueva al carrousel.');
     }
 
     /**
@@ -38,9 +40,12 @@ class FotoController extends Controller
      */
     public function destroy(Foto $foto)
     {
+        if (! Gate::allows('es_propietario-foto', $foto)) {
+            abort(403);
+        }
         $evento=$foto->evento;
         Storage::delete($foto->ruta_foto);
         $foto->delete();
-        return redirect()->route('evento.edit',$evento);
+        return redirect()->route('evento.edit',$evento)->with('success','Se elimino una foto del carrousel.');
     }
 }
