+++ app/Http/Controllers/InvitadoController.php
@@ -8,6 +8,7 @@ use App\Models\Invitado;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Gate;
 
 class InvitadoController extends Controller
 {
@@ -18,6 +19,9 @@ class InvitadoController extends Controller
      */
     public function index(Evento $evento)
     {
+        if (! Gate::allows('es_propietario-evento', $evento)) {
+            abort(403);
+        }
         Auth::user()->setEventoActual($evento);
         $invitados=$evento->invitados;
         return view('invitado/invitadoIndex',compact('invitados','evento'));
@@ -51,7 +55,7 @@ class InvitadoController extends Controller
 
         $invitado=Invitado::create($request->all());
         $evento=$invitado->evento;
-        return redirect()->route('index_invitados',$evento);
+        return redirect()->route('index_invitados',$evento)->with('success','Se creo el usuario correctamente');
     }
 
     /**
@@ -84,12 +88,16 @@ class InvitadoController extends Controller
      */
     public function update(Request $request, Invitado $invitado)
     {
+        if (! Gate::allows('es_propietario-invitado', $invitado)) {
+            abort(403);
+        }
         $request->validate([
             'confirmacion' => 'required'
 
         ]);
-        Invitado::where('id',$invitado->id)->update($request->except('_token','_method'));
-        return redirect()->back()->with('success','Se ha enviado la confirmacion');
+        $invitado->confirmacion=$request->confirmacion;
+        $invitado->save();
+        return redirect()->back()->with('success','Se ha enviado la confirmacion.');
     }
 
     /**
@@ -100,12 +108,19 @@ class InvitadoController extends Controller
      */
     public function destroy(Invitado $invitado)
     {
+        if (! Gate::allows('es_propietario-invitado', $invitado)) {
+            abort(403);
+        }
         $evento=$invitado->evento;
         $invitado->delete();
-        return redirect()->route('index_invitados',$evento);
+        return redirect()->route('index_invitados',$evento)->with('success','Se elimino el usuario correctamente.');
     }
     public function enviarInvitacion(Evento $evento, Invitado $invitado){
+        if (! Gate::allows('es_propietario-invitado', $invitado)) {
+            abort(403);
+        }
         mail::to($invitado->correo_invitado)->send(new invitacionMail($evento,$invitado));
-        return redirect()->back();
+        return redirect()->back()->with('success','Se han enviado el correo correctamente.');
     }
+
 }
