+++ app/Http/Controllers/MesaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Mesa;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class MesaController extends Controller
 {
@@ -18,13 +19,16 @@ class MesaController extends Controller
 
         $mesa=Mesa::create($request->all());
         $evento=$mesa->evento;
-        return redirect()->route('evento.edit',$evento);
+        return redirect()->route('evento.edit',$evento)->with('success','Se ha agregado la mesa de regalos.');
     }
     public function destroy(Mesa $mesa)
     {
+        if (! Gate::allows('es_propietario-mesa', $mesa)) {
+            abort(403);
+        }
         $evento=$mesa->evento;
         $mesa->delete();
-        return redirect()->route('evento.edit',$evento);
+        return redirect()->route('evento.edit',$evento)->with('success','Se ha eliminado una mesa de regalos.');
     }
 
 }
